Splitting right dock into plot tree and styling docks
The right dock has been split in two: a Plot Tree dock and a Styling dock.  The plot type selection and plot update button have been moved to the main toolbar.  This allows a bit more flexibility for the user without much code refactoring.  All the widgets work the same as before despite their new locations.  The only widget removed was a toolbox that contained the Plot Selector widgets and Styling toolbox, the contents of each page of which became the new docks.  dockWidgetRightToolbox is now dockWidgetPlotTree.