Move field type and field selection outside control toolbox
The field type and field selection within Control Toolbox and ColorByField and ColorField in Styling Toolbox>Colors have been moved to the a group box in the Control Dock.  This should reduce the number properties to keep track of.  To do so, has required changing MainWindow.plot_types to Main_Window.plot dict, which holds settings for the field type/field comboboxes associated with each change in MainWindow.toolBox.curretnIndex().