More Logger related updates
I fixed several places where errors related to the logger occur.  They often seem to happen where a signal is sent from a widget by a connection to a function that does not take any arguments.  In this case, *args and **kwargs should be added as function inputs that will be absorbed and promptly ignored when the function is called.