Updated ternary colors and started on new functionality in NotesDock
Ternary colormaps can now be loaded, altered and saved.  They are not applied to maps yet.

Notes are being refactored to receive data of various types and convert them to formatted reSTructured text for display in the NotesDock.text_edit.  This should help with making Notes modular, independent of LaME.