Completed ColormapEditor and associated widgets+dialogs
The ColormapEditor is now finished aside from any bugs.  It can be run standalone as a dialog or incorporated into a program.  It also includes a ColormapSelector widget that can be used separately to choose a colormap.  There is also a new ColorManager that includes a lot of helper functions for dealing with converting colors between the various systems (HSV, RGB, CMYK, HEX) and includes QColor for PyQt applications.